ksmbd-tools: mountd: add sock_listener
This module defines a netlink socket to receive notifications about
interfaces state changes.
At initialization, a RTM_GETLINK message is sent to receive the current
state of all interfaces.
Responses are processed according to the configuration: if the
'bind_interfaces_only' parameter is not set, a TCP listener is created.
Otherwise, the list of allowed interfaces is checked first.

The listener sockets are polled: if a client initiates a connection, it
is accept()'ed and then passed to the ksmbd kernel module via netlink.

This lets ksmbd listen for connections on interfaces that come up
after startup.

Additionally, since the listening sockets are handled in userspace,
restarting ksmbd.mountd with an updated configuration file is enough to
change the list of interfaces to listen on (no need to stop the kernel
module and interrupt established sessions).